pyproject.toml
[project]
name = "backend-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastmcp>=2.13.0.2",
    "sqlalchemy",
    "psycopg2-binary",
    "alembic",
    "python-dotenv",
]
